requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.5.4
djangorestframework-gis==0.11
elasticsearch==5.3.0
olefile==0.44
packaging==16.8
Pillow-SIMD==4.0.0.post0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
urllib3==1.21
